refactor navbar to DRY and keep naming semantic
