+++ src/frontend/src/components/NavBar/components/AppNavBar/components/GroupDetailsDropdown/components/GroupMenuItem/style.ts
