+++ src/frontend/src/components/NavBar/components/AppNavBar/components/GroupDetailsDropdown/index.tsx
