+++ src/frontend/src/components/NavBar/components/AppNavBar/components/GroupDetailsDropdown/style.ts
