+++ src/frontend/src/components/NavBar/components/AppNavBar/style.ts
