+++ src/frontend/src/components/NavBar/components/StaticPageNavBar/style.ts
